Revert to js
