+++ package.json
@@ -16,6 +16,7 @@
     "googleapis": "^83.0.0",
     "http": "0.0.0",
     "node-persist": "^2.1.0",
+    "openai": "^2.0.5",
     "showdown": "^1.9.1",
     "turndown": "^6.0.0"
   },
